# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Job {
  id: number;
  movieTitle: string;
  language: string;
  status: JobStatus;
  progress: number;
  stage: string;
  outputUrl?: string | null;
  thumbnailUrl?: string | null;
  durationSeconds?: number | null;
  recommendedDuration?: number | null;
  error?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface CreateJobBody {
  movieTitle: string;
  language: string;
  videoFilename?: string | null;
}

export type UpdateJobProgressBodyStatus = typeof UpdateJobProgressBodyStatus[keyof typeof UpdateJobProgressBodyStatus];


export const UpdateJobProgressBodyStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface UpdateJobProgressBody {
  status?: UpdateJobProgressBodyStatus;
  progress?: number;
  stage?: string;
  outputUrl?: string | null;
  thumbnailUrl?: string | null;
  durationSeconds?: number | null;
  recommendedDuration?: number | null;
  error?: string | null;
}

export interface JobStats {
  total: number;
  completed: number;
  processing: number;
  failed: number;
  pending: number;
}

export interface GetJobParams {
  id: number;
}

export interface DeleteJobParams {
  id: number;
}

export interface UpdateJobProgressParams {
  id: number;
}