# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary List all jobs
 */
export const ListJobsResponseItem = zod.object({
  "id": zod.number(),
  "movieTitle": zod.string(),
  "language": zod.string(),
  "status": zod.enum(['pending', 'processing', 'completed', 'failed']),
  "progress": zod.number(),
  "stage": zod.string(),
  "outputUrl": zod.string().nullish(),
  "thumbnailUrl": zod.string().nullish(),
  "durationSeconds": zod.number().nullish(),
  "recommendedDuration": zod.number().nullish(),
  "error": zod.string().nullish(),
  "createdAt": zod.coerce.date(),
  "completedAt": zod.coerce.date().nullish()
})
export const ListJobsResponse = zod.array(ListJobsResponseItem)


/**
 * @summary Create a new job
 */
export const CreateJobBody = zod.object({
  "movieTitle": zod.string(),
  "language": zod.string(),
  "videoFilename": zod.string().nullish()
})


/**
 * @summary Get job statistics
 */
export const GetJobStatsResponse = zod.object({
  "total": zod.number(),
  "completed": zod.number(),
  "processing": zod.number(),
  "failed": zod.number(),
  "pending": zod.number()
})


/**
 * @summary Get a single job
 */
export const GetJobParams = zod.object({
  "id": zod.coerce.number()
})

export const GetJobResponse = zod.object({
  "id": zod.number(),
  "movieTitle": zod.string(),
  "language": zod.string(),
  "status": zod.enum(['pending', 'processing', 'completed', 'failed']),
  "progress": zod.number(),
  "stage": zod.string(),
  "outputUrl": zod.string().nullish(),
  "thumbnailUrl": zod.string().nullish(),
  "durationSeconds": zod.number().nullish(),
  "recommendedDuration": zod.number().nullish(),
  "error": zod.string().nullish(),
  "createdAt": zod.coerce.date(),
  "completedAt": zod.coerce.date().nullish()
})


/**
 * @summary Delete a job
 */
export const DeleteJobParams = zod.object({
  "id": zod.coerce.number()
})


/**
 * @summary Update job progress
 */
export const UpdateJobProgressParams = zod.object({
  "id": zod.coerce.number()
})

export const UpdateJobProgressBody = zod.object({
  "status": zod.enum(['pending', 'processing', 'completed', 'failed']).optional(),
  "progress": zod.number().optional(),
  "stage": zod.string().optional(),
  "outputUrl": zod.string().nullish(),
  "thumbnailUrl": zod.string().nullish(),
  "durationSeconds": zod.number().nullish(),
  "recommendedDuration": zod.number().nullish(),
  "error": zod.string().nullish()
})

export const UpdateJobProgressResponse = zod.object({
  "id": zod.number(),
  "movieTitle": zod.string(),
  "language": zod.string(),
  "status": zod.enum(['pending', 'processing', 'completed', 'failed']),
  "progress": zod.number(),
  "stage": zod.string(),
  "outputUrl": zod.string().nullish(),
  "thumbnailUrl": zod.string().nullish(),
  "durationSeconds": zod.number().nullish(),
  "recommendedDuration": zod.number().nullish(),
  "error": zod.string().nullish(),
  "createdAt": zod.coerce.date(),
  "completedAt": zod.coerce.date().nullish()
})