# lib/api-zod/src/generated/types/createJobBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateJobBody {
  movieTitle: string;
  language: string;
  videoFilename?: string | null;
}