# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';

export interface Job {
  id: number;
  movieTitle: string;
  language: string;
  status: JobStatus;
  progress: number;
  stage: string;
  outputUrl?: string | null;
  thumbnailUrl?: string | null;
  durationSeconds?: number | null;
  recommendedDuration?: number | null;
  error?: string | null;
  createdAt: Date;
  completedAt?: Date | null;
}