# lib/api-zod/src/generated/types/jobStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobStats {
  total: number;
  completed: number;
  processing: number;
  failed: number;
  pending: number;
}