# lib/api-zod/src/generated/types/updateJobProgressBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateJobProgressBodyStatus } from './updateJobProgressBodyStatus';

export interface UpdateJobProgressBody {
  status?: UpdateJobProgressBodyStatus;
  progress?: number;
  stage?: string;
  outputUrl?: string | null;
  thumbnailUrl?: string | null;
  durationSeconds?: number | null;
  recommendedDuration?: number | null;
  error?: string | null;
}